requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
blinker==1.4
celery==5.4.0
cliapp==1.20180812.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmdtest==0.32+git
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.3.0
Django==5.0.6
django-axes==7.0.1
django-cors-headers==4.4.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==1.1.2
gunicorn==23.0.0
gyp==0.1
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kombu==5.4.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pynvim==0.4.2
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
redis==5.2.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.0
systemd-python==234
transitions==0.9.2
ttystatus==0.38
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
whitenoise==6.9.0
zipp==1.0.0
